Add files via upload
CNN  is  the  most  popular  deep  learning  algorithm,  used tremendously   in   computer   vision.   It   is   computationally efficient  and  instinctively  spots  prominent  features  without any supervision. We have implemented an efficient method in the detection of gender   and age.   In comparison   to   other   deep   learning approaches, the identification procedure was simple, took less time  to compute, and used less memory with our CNN. The resulting results are precise and straightforward, with accurate predictions; the person's gender is appropriately predicted as male or female. Due to the application of bias, the algorithm produces precise results when considering age.